requirements.txt
# Core dependencies
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0

# Database dependencies
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

# Environment configuration
python-dotenv==1.0.1

# Airflow (only needed if running Airflow locally, not in Docker)
# apache-airflow==2.8.1
